+++ test/index.js
@@ -1,7 +1,9 @@
 require('./bem-pattern');
 require('./definition');
+require('./ignore-custom-properties');
 require('./ignore-selectors');
 require('./property-validation');
+require('./ranges');
 require('./selector-validation');
 require('./suit-pattern');
 require('./utility-validation');